+++ internal/impl/couchbase/integration_test.go
@@ -2,15 +2,20 @@ package couchbase_test
 
 import (
 	"bytes"
+	"context"
 	"fmt"
 	"os"
 	"sync"
 	"testing"
 	"time"
 
+	"github.com/go-faker/faker/v4"
 	"github.com/ory/dockertest/v3"
 	"github.com/ory/dockertest/v3/docker"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"github.com/warpstreamlabs/bento/public/service"
+	"github.com/warpstreamlabs/bento/public/service/integration"
 )
 
 var (
@@ -105,4 +110,255 @@ func setupCouchbase(tb testing.TB) (*dockertest.Pool, *dockertest.Resource, erro
 	tb.Log("couchbase cluster ready")
 
 	return pool, resource, nil
+}
+
+func TestIntegrationCouchbaseProcessor(t *testing.T) {
+	integration.CheckSkip(t)
+
+	servicePort := requireCouchbase(t)
+
+	bucket := fmt.Sprintf("testing-processor-%d", time.Now().Unix())
+	require.NoError(t, createBucket(context.Background(), t, servicePort, bucket))
+	t.Cleanup(func() {
+		require.NoError(t, removeBucket(context.Background(), t, servicePort, bucket))
+	})
+
+	uid := faker.UUIDHyphenated()
+	payload := fmt.Sprintf(`{"id": %q, "data": %q}`, uid, faker.Sentence())
+
+	t.Run("Insert", func(t *testing.T) {
+		testCouchbaseProcessorInsert(uid, payload, bucket, servicePort, t)
+	})
+	t.Run("Get", func(t *testing.T) {
+		testCouchbaseProcessorGet(uid, payload, bucket, servicePort, t)
+	})
+	t.Run("Remove", func(t *testing.T) {
+		testCouchbaseProcessorRemove(uid, bucket, servicePort, t)
+	})
+	t.Run("GetMissing", func(t *testing.T) {
+		testCouchbaseProcessorGetMissing(uid, bucket, servicePort, t)
+	})
+
+	payload = fmt.Sprintf(`{"id": %q, "data": %q}`, uid, faker.Sentence())
+	t.Run("Upsert", func(t *testing.T) {
+		testCouchbaseProcessorUpsert(uid, payload, bucket, servicePort, t)
+	})
+	t.Run("Get", func(t *testing.T) {
+		testCouchbaseProcessorGet(uid, payload, bucket, servicePort, t)
+	})
+
+	payload = fmt.Sprintf(`{"id": %q, "data": %q}`, uid, faker.Sentence())
+	t.Run("Replace", func(t *testing.T) {
+		testCouchbaseProcessorReplace(uid, payload, bucket, servicePort, t)
+	})
+	t.Run("Get", func(t *testing.T) {
+		testCouchbaseProcessorGet(uid, payload, bucket, servicePort, t)
+	})
+}
+
+func TestIntegrationCouchbaseStream(t *testing.T) {
+	ctx := context.Background()
+
+	integration.CheckSkip(t)
+
+	servicePort := requireCouchbase(t)
+	bucket := fmt.Sprintf("testing-stream-%d", time.Now().Unix())
+	require.NoError(t, createBucket(context.Background(), t, servicePort, bucket))
+	t.Cleanup(func() {
+		require.NoError(t, removeBucket(context.Background(), t, servicePort, bucket))
+	})
+
+	for _, clearCAS := range []bool{true, false} {
+		t.Run(fmt.Sprintf("%t", clearCAS), func(t *testing.T) {
+			streamOutBuilder := service.NewStreamBuilder()
+			require.NoError(t, streamOutBuilder.SetLoggerYAML(`level: OFF`))
+
+			inFn, err := streamOutBuilder.AddBatchProducerFunc()
+			require.NoError(t, err)
+
+			var outBatches []service.MessageBatch
+			var outBatchMut sync.Mutex
+			require.NoError(t, streamOutBuilder.AddBatchConsumerFunc(func(c context.Context, mb service.MessageBatch) error {
+				outBatchMut.Lock()
+				outBatches = append(outBatches, mb)
+				outBatchMut.Unlock()
+				return nil
+			}))
+
+			// insert
+			require.NoError(t, streamOutBuilder.AddProcessorYAML(fmt.Sprintf(`
+couchbase:
+  url: 'couchbase://localhost:%s'
+  bucket: %s
+  username: %s
+  password: %s
+  id: '${! json("key") }'
+  content: 'root = this'
+  operation: 'insert'
+`, servicePort, bucket, username, password)))
+
+			if clearCAS { // ignore cas check
+				require.NoError(t, streamOutBuilder.AddProcessorYAML(`
+mapping: |
+  meta couchbase_cas = deleted()
+`))
+			}
+
+			// replace
+			require.NoError(t, streamOutBuilder.AddProcessorYAML(fmt.Sprintf(`
+couchbase:
+  url: 'couchbase://localhost:%s'
+  bucket: %s
+  username: %s
+  password: %s
+  id: '${! json("key") }'
+  content: 'root = this'
+  operation: 'replace'
+`, servicePort, bucket, username, password)))
+
+			if clearCAS { // ignore cas check
+				require.NoError(t, streamOutBuilder.AddProcessorYAML(`
+mapping: |
+  meta couchbase_cas = deleted()
+`))
+			}
+			// remove
+			require.NoError(t, streamOutBuilder.AddProcessorYAML(fmt.Sprintf(`
+couchbase:
+  url: 'couchbase://localhost:%s'
+  bucket: %s
+  username: %s
+  password: %s
+  id: '${! json("key") }'
+  operation: 'remove'
+`, servicePort, bucket, username, password)))
+
+			streamOut, err := streamOutBuilder.Build()
+			require.NoError(t, err)
+			go func() {
+				err = streamOut.Run(context.Background())
+				require.NoError(t, err)
+			}()
+
+			require.NoError(t, inFn(ctx, service.MessageBatch{
+				service.NewMessage([]byte(`{"key":"hello","value":"word"}`)),
+			}))
+			require.NoError(t, streamOut.StopWithin(time.Second*15))
+
+			assert.Eventually(t, func() bool {
+				outBatchMut.Lock()
+				defer outBatchMut.Unlock()
+				return len(outBatches) == 1
+			}, time.Second*5, time.Millisecond*100)
+
+			// batch processing should be fine and contain one message.
+			assert.NoError(t, err)
+			assert.Len(t, outBatches, 1)
+			assert.Len(t, outBatches[0], 1)
+
+			// message should contain an error.
+			assert.NoError(t, outBatches[0][0].GetError())
+		})
+	}
+}
+
+func TestIntegrationCouchbaseStreamError(t *testing.T) {
+	ctx := context.Background()
+
+	integration.CheckSkip(t)
+
+	servicePort := requireCouchbase(t)
+	bucket := fmt.Sprintf("testing-stream-error-%d", time.Now().Unix())
+	require.NoError(t, createBucket(context.Background(), t, servicePort, bucket))
+	t.Cleanup(func() {
+		require.NoError(t, removeBucket(context.Background(), t, servicePort, bucket))
+	})
+
+	streamOutBuilder := service.NewStreamBuilder()
+	require.NoError(t, streamOutBuilder.SetLoggerYAML(`level: OFF`))
+
+	inFn, err := streamOutBuilder.AddBatchProducerFunc()
+	require.NoError(t, err)
+
+	var outBatches []service.MessageBatch
+	var outBatchMut sync.Mutex
+	require.NoError(t, streamOutBuilder.AddBatchConsumerFunc(func(c context.Context, mb service.MessageBatch) error {
+		outBatchMut.Lock()
+		outBatches = append(outBatches, mb)
+		outBatchMut.Unlock()
+		return nil
+	}))
+
+	// insert
+	require.NoError(t, streamOutBuilder.AddProcessorYAML(fmt.Sprintf(`
+couchbase:
+  url: 'couchbase://localhost:%s'
+  bucket: %s
+  username: %s
+  password: %s
+  id: '${! json("key") }'
+  content: |
+    root = this
+    root.at = timestamp_unix_micro()
+  operation: 'insert'
+`, servicePort, bucket, username, password)))
+
+	// upsert and remove in parallel
+	require.NoError(t, streamOutBuilder.AddProcessorYAML(fmt.Sprintf(`
+workflow:
+  meta_path: ""
+  branches:
+    write:
+      processors:
+        - couchbase:
+            url: 'couchbase://localhost:%[1]s'
+            bucket: %[2]s
+            username: %[3]s
+            password: %[4]s
+            id: '${! json("key") }'
+            content: |
+              root = this
+              root.at = timestamp_unix_micro()
+            operation: 'replace'
+    remove:
+      processors:
+        - sleep:
+            duration: "1s"
+        - couchbase:
+            url: 'couchbase://localhost:%[1]s'
+            bucket: %[2]s
+            username: %[3]s
+            password: %[4]s
+            id: '${! json("key") }'
+            content: |
+              root = this
+              root.at = timestamp_unix_micro()
+            operation: 'replace'
+`, servicePort, bucket, username, password)))
+
+	streamOut, err := streamOutBuilder.Build()
+	require.NoError(t, err)
+	go func() {
+		err = streamOut.Run(context.Background())
+		require.NoError(t, err)
+	}()
+
+	require.NoError(t, inFn(ctx, service.MessageBatch{
+		service.NewMessage([]byte(`{"key":"hello","value":"word"}`)),
+	}))
+	require.NoError(t, streamOut.StopWithin(time.Second*15))
+
+	assert.Eventually(t, func() bool {
+		outBatchMut.Lock()
+		defer outBatchMut.Unlock()
+		return len(outBatches) == 1
+	}, time.Second*5, time.Millisecond*100)
+
+	// batch contain one message.
+	assert.NoError(t, err)
+	assert.Len(t, outBatches, 1)
+	assert.Len(t, outBatches[0], 1)
+
+	// message should contain an error.
+	assert.Error(t, outBatches[0][0].GetError())
 }
